# cmd/am2server/openapi.yml
openapi: 3.0.3
info:
  contact:
      name: sqlc-http
      url: https://github.com/walterwanderley/sqlc-http
  description: Manage and share am2/am2data (SincoANN profile) files
  title: am2 Server
  version: 0.0.1
...
paths:
...
  /reviews:
      post:
          requestBody:
              content:
                  application/json:
                      schema:
                          properties:
                              capture_id:
                                  format: int64
                                  type: integer
                              comment:
                                  type: string
                              rate:
                                  format: int64
                                  type: integer
                              user_id:
                                  format: int64
                                  type: integer
                          type: object
                  application/x-www-form-urlencoded:
                      schema:
                          properties:
                              capture_id:
                                  format: int64
                                  type: integer
                              comment:
                                  type: string
                              rate:
                                  format: int64
                                  type: integer
                              user_id:
                                  format: int64
                                  type: integer
                          type: object
          responses:
              "200":
                  content:
                      application/json:
                          schema:
                              properties:
                                  last_insert_id:
                                      format: int64
                                      type: integer
                                  rows_affected:
                                      format: int64
                                      type: integer
                              type: object
                  description: OK
              default:
                  content:
                      text/plain:
                          schema:
                              type: string
                  description: Error message
          summary: AddReview
          tags:
              - review
  /reviews/{id}:
      delete:
          parameters:
              - in: path
                name: id
                schema:
                  format: int64
                  type: integer
          responses:
              "200":
                  content:
                      application/json:
                          schema:
                              properties:
                                  last_insert_id:
                                      format: int64
                                      type: integer
                                  rows_affected:
                                      format: int64
                                      type: integer
                              type: object
                  description: OK
              default:
                  content:
                      text/plain:
                          schema:
                              type: string
                  description: Error message
          summary: RemoveReview
          tags:
              - review
  /users/{id}:
      get:
          parameters:
              - in: path
                name: id
                schema:
                  format: int64
                  type: integer
          responses:
              "200":
                  content:
                      application/json:
                          schema:
                              $ref: '#/components/schemas/userUser'
                  description: OK
              default:
                  content:
                      text/plain:
                          schema:
                              type: string
                  description: Error message
          summary: GetUser
          tags:
              - user
  /users/{id}/name:
      patch:
          parameters:
              - in: path
                name: id
                schema:
                  format: int64
                  type: integer
          requestBody:
              content:
                  application/json:
                      schema:
                          properties:
                              name:
                                  type: string
                          type: object
                  application/x-www-form-urlencoded:
                      schema:
                          properties:
                              name:
                                  type: string
                          type: object
          responses:
              "200":
                  content:
                      application/json:
                          schema:
                              properties:
                                  last_insert_id:
                                      format: int64
                                      type: integer
                                  rows_affected:
                                      format: int64
                                      type: integer
                              type: object
                  description: OK
              default:
                  content:
                      text/plain:
                          schema:
                              type: string
                  description: Error message
          summary: UpdateUserName
          tags:
              - user
...
components:
  schemas:
...
    userUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        picture:
          type: string